# src/main/resources/api.yml
openapi: 3.0.0
info:
  title: Campus Wall API
  version: "1.0.0"
  description: API for campus social features
servers:
  - url: /api/v1
...
paths:
...
  /marketplace/{itemId}:
    get:
      tags: [marketplace]
      summary: Get marketplace item by ID
      description: Retrieve a specific marketplace item by its ID
      operationId: getItem
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
          description: ID of the marketplace item to retrieve
      responses:
        "200":
          description: Item retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDTO"
        "401":
          description: Unauthorized - authentication required
        "404":
          description: Item not found
    put:
      tags: [marketplace]
      summary: Update marketplace item
      description: Update an existing marketplace item (partial update supported)
      operationId: updateItem
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
          description: ID of the marketplace item to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
      responses:
        "200":
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDTO"
        "400":
          description: Invalid request data
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - not the owner of the item
        "404":
          description: Item not found
...
  /admin/marketplaces:
    get:
      tags: [admin]
      summary: Get all marketplace items
      operationId: adminGetMarketplaces
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: hidden
          schema:
            type: boolean
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [createdAt, price, commentCount]
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Marketplace items retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminMarketplaceDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/marketplaces/{id}:
    get:
      tags: [admin]
      summary: Get a marketplace item by ID
      operationId: adminGetMarketplaceById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Marketplace item retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMarketplaceDTO"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/marketplaces/{id}/hide:
    put:
      tags: [admin]
      summary: Hide a marketplace item
      operationId: adminHideMarketplace
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Marketplace item hidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/marketplaces/{id}/unhide:
    put:
      tags: [admin]
      summary: Unhide a marketplace item
      operationId: adminUnhideMarketplace
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Marketplace item unhidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
components:
  schemas:
...
    UpdateItemRequest:
      type: object
      description: Request body for updating a marketplace item. All fields are optional for partial updates.
      properties:
        title:
          type: string
          description: Item title
        description:
          type: string
          description: Item description
        price:
          type: number
          format: float
          description: Item price
        category:
          type: string
          description: Item category (e.g., books, electronics, furniture)
        condition:
          type: string
          enum: [new, like-new, good, fair]
          description: Condition of the item
        sold:
          type: boolean
          description: Whether the item has been sold
    ItemDTO:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
        condition:
          type: string
          enum: [new, like-new, good, fair]
        sold:
          type: boolean
          default: false
        wall:
          type: string
          enum: [campus, national]
          description: Which wall the item is on
        comments:
          type: integer
          description: Number of comments
        author:
          type: object
          properties:
            id:
              type: string
            profileName:
              type: string
              description: Profile name of the item author
            isAnonymous:
              type: boolean
              description: Whether author is anonymous
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ----- Coins -----
...
    AdminMarketplaceDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        profileName:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        condition:
          type: string
        sold:
          type: boolean
        wall:
          type: string
          enum: [campus, national]
        schoolDomain:
          type: string
        commentCount:
          type: integer
        hidden:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT